+++ app/controllers/concerns/manageable.rb
@@ -3,6 +3,7 @@ module Concerns
     extend ActiveSupport::Concern
 
     def create
+      authorize model_class
       display_type = params[:"#{model_type}"][:display_type]
       form = model_class::Form.new(model_class.new)
 
@@ -15,6 +16,7 @@ module Concerns
     end
 
     def update
+      authorize model_class
       display_type = params[:"#{model_type}"][:display_type]
       model = model_class.find(params[:id])
       form = model_class::Form.new(model)
@@ -28,6 +30,7 @@ module Concerns
     end
 
     def destroy
+      authorize model_class
       model = model_class.find(params[:id])
       model.destroy
 
@@ -40,22 +43,26 @@ module Concerns
     end
 
     def show_new_form
+      authorize model_class
       form = model_class::Form.new(model_class.new)
       render js: concept("#{model_type}/#{model_type}_form_cell", form, current_user: current_user, display_type: params[:display_type]).(:show_new_form)
     end
 
     def cancel_new_form
+      authorize model_class
       form = model_class::Form.new(model_class.new)
       render js: concept("#{model_type}/#{model_type}_form_cell", form, current_user: current_user).(:cancel_new_form)
     end
 
     def show_edit_form
+      authorize model_class
       model = model_class.find(params[:"#{model_type}_id"])
       form = model_class::Form.new(model)
       render js: concept("#{model_type}/#{model_type}_form_cell", form, current_user: current_user, display_type: params[:display_type]).(:show_edit_form)
     end
 
     def cancel_edit_form
+      authorize model_class
       display_type = params[:display_type]
       model = model_class.find(params[:"#{model_type}_id"])
       render js: concept("#{model_type}/#{model_type}_cell/#{display_type}", model, current_user: current_user).(:refresh)
